+++ app.js
@@ -15,6 +15,12 @@ app.use(express.json());
 
 app.use('/api/contacts', contactsRouter);
 
+app.all("*", (req, res) => {
+  res.status(404).json({
+    mesages: "Oooops! Resource not found!",
+  });
+});
+
 app.use((req, res) => {
   res.status(404).json({ message: 'Not found' });
 });
